add membership

+++ app/Services/GroupService.php
@@ -44,6 +44,8 @@ class GroupService extends Service
             'name' => Str::slug($request['name'], '-'),
             'type' => $request['type'],
             'description' => $request['description'],
+            'bg_image_url' => null,
+            'icon_image_url'=> null
            ];
         if($request->hasFile('bg_image')){
             $data['bg_image_url'] = Storage::put($path, $request['bg_image']);
@@ -52,36 +54,67 @@ class GroupService extends Service
             $data['icon_image_url'] = Storage::put($path, $request['icon_image']);
         }  
         
-
-
         $group = $user->groups()->create($data);
-        $data['id'] = $group->id;
-        $data['bg_image_url'] = $group['bg_image_url'];
-        $data['icon_image_url'] = $group['bg_image_url'];
         return $this->responseService->message('The group has been created successfully')
-            ->status(201)->data($data);
+            ->status(201)->data($group);
 
 
     }
-    public function update(Request $request,Group $group)
-    {
-        //
-    }
+
     public function show(Group $group)
     {
-        
-        $group['bg_image_url']= Storage::temporaryUrl(
-            $group['bg_image_url'], now()->addMinutes(5)
-        );
-        $group['icon_image_url']= Storage::temporaryUrl(
-            $group['icon_image_url'], now()->addMinutes(5)
-        );
-        return $this->responseService->data($group);
+        $user = User::find(Auth::id());
+        if ($user->cannot('view', $group)) {
+            return $this->responseService->message('unauthorized')
+                ->status(403)->error(true);
+        }
+        $data = [
+            'group' => $group ,
+            'files' => $group->files()->get()
+        ];
+        return $this->responseService->data($data);
+    }
+    public function update(Request $request ,Group $group)
+    {
+        $user = User::find(Auth::id());
+        if ($user->cannot('update', $group)) {
+            return $this->responseService->message('unauthorized')
+                ->status(403)->error(true);
+        }
+        $errors = Validator::make($request->all(),['description'=>['nullable','string', 'max:2048']])->errors()->all();
+        if ($errors) {
+            return $this->responseService->message($errors)->status(404)->error(true);
+        }
+        if($request->has('description')){
+            $group['description'] = $request['description'];
+        }
+        $path = 'images/'.$user['user_name'].'/'. $group['name'];
+        if($request->hasFile('bg_image')){
+            if($group->bgImageOriginalUrl)
+                Storage::delete($group->bgImageOriginalUrl);
+            $group['bg_image_url'] = Storage::put($path, $request['bg_image']);
+        }
+        if($request->hasFile('icon_image')){
+            if($group->iconImageOriginalUrl)
+                Storage::delete($group->iconImageOriginalUrl);
+            $group['icon_image_url'] = Storage::put($path, $request['icon_image']);
+        } 
+
+        $group->save();
+        return $this->responseService->message('The group has been updated successfully')
+            ->data($group)->status(201);
     }
     public function destroy(Group $group)
     {
+        $user = User::find(Auth::id());
+        if ($user->cannot('delete', $group)) {
+            return $this->responseService->message('unauthorized')
+                ->status(403)->error(true);
+        }
+        Storage::deleteDirectory($group->filesDirectory);
+        Storage::deleteDirectory($group->imagesDirectory);
         $group->delete();
-        return $this->responseService->message('The group has been created successfully')
+        return $this->responseService->message('The group has been deleted successfully')
             ->status(204);
     }
